update show tests
Julia 1.6 has made a lot of changes to `show`. Two of the most "breaking
change" among it is 1) that `Array{T, 2}` is now shown as `Matrix{T}` and
2) that `RGB{Normed{UInt8, 8}}` is now shown as its alias `RGB{N0f8}`

FixedPointNumbers and Colors/ColorTypes have made similar changes, too.

This update is not perfect because the reference is hard coded into test cases
and it's very likely to break in future. But given that how `op` is shown isn't
of high priority here, here I just save myself some time and copy & paste the
results so as to make tests pass.